openSUSE: do not run random stuff off the internet and run it
The build_image.sh is already needed to run with super user privileges
for block device access. Definitivley not a great idea to download
any shell script off the internet and run it as superuser

If we want shells-helper in the Shells.com images, we have to package
shells-helper and install it as a package, as anything else is done.